# services/control-plane-api/openapi/control_plane_openapi_v1.4.yaml
openapi: 3.0.3
info:
  title: AI Digital Office OS Control Plane
  version: 1.4.0
  description: >
    Governed control-plane API for AI DIGITAL OFFICE OS. Every endpoint is
    tenant-scoped via the JWT 'tenant_id' claim, versioned under /v1, and
    returns the standard error envelope on any non-2xx response (see
    ErrorResponse schema). v1.4 closes the integrity-review findings against
    v1.3: full project/agent-version/run/deployment/usage/tool lifecycle
    coverage, an explicit auth contract (issuer/audience/JWKS/scopes), and
    a required Idempotency-Key header on every state-changing endpoint.
    This specification is product-agnostic: no endpoint, schema, or example
    is specific to any named downstream project (see blueprint clause 75).
servers:
  - url: https://api.digitaloffice.example/v1
    description: Production (placeholder host — real endpoint is an
      environment-specific choice, intentionally not fabricated here)
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        OAuth2 client-credentials token for service/agent callers, or
        OAuth2 authorization-code token for human users (blueprint clause 70).
        Required claims: 'iss' (issuer — this Office's or an approved
        external IdP), 'aud' (this control plane), 'exp'/'iat' (max lifetime
        15 minutes for service/agent callers, 8 hours for human sessions),
        'tenant_id', 'scope' (granted permission scopes), and 'actor_type'
        (SERVICE | HUMAN). Clock-skew tolerance ±60 seconds. JWKS rotates on
        a maximum 90-day cycle with a minimum 7-day overlap window. Session
        and refresh-token lifecycle is delegated entirely to the issuing IdP
        (blueprint clause 71) — this API only validates presented tokens.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string }
      description: >
        Required on every state-changing (POST/PATCH/DELETE) request.
        Replaying the same key within 24 hours returns the original result
        rather than reprocessing (blueprint clauses 49, 63, 68).
  schemas:
    ErrorResponse:
      type: object
      required: [error_code, message, correlation_id, retryable]
      properties:
        error_code:
          type: string
          enum: [VALIDATION_ERROR, AUTHORIZATION_ERROR, POLICY_ERROR, TOOL_ERROR,
                 MODEL_ERROR, PROVIDER_ERROR, NETWORK_ERROR, TIMEOUT,
                 DATABASE_ERROR, ARTIFACT_ERROR, DEPLOYMENT_ERROR, UNKNOWN_ERROR]
        message: { type: string }
        correlation_id: { type: string }
        retryable: { type: boolean }
    Task:
      type: object
      required: [task_id, tenant_id, status]
      properties:
        task_id: { type: string }
        tenant_id: { type: string, format: uuid }
        project_id: { type: string }
        workflow_id: { type: string }
        parent_task_id: { type: string }
        assigned_agent: { type: string }
        required_capability: { type: string }
        priority: { type: string }
        risk_level: { type: string, enum: [GREEN, YELLOW, RED] }
        security_level: { type: string }
        dependencies: { type: array, items: { type: string } }
        input: { type: object }
        expected_output: { type: object }
        status:
          type: string
          enum: [CREATED, QUEUED, ASSIGNED, RUNNING, WAITING, BLOCKED, FAILED, RETRYING, COMPLETED, CANCELLED, ESCALATED]
        retry_count: { type: integer }
        idempotency_key: { type: string }
        deadline: { type: string, format: date-time }
...
    Workflow:
      type: object
      properties:
        workflow_id: { type: string }
        tenant_id: { type: string, format: uuid }
        project_id: { type: string }
        workflow_type: { type: string }
        definition_version: { type: string }
        current_state: { type: object }
        status: { type: string, enum: [RUNNING, PAUSED, COMPLETED, FAILED, CANCELLED, ESCALATED] }
        started_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
    Artifact:
      type: object
      properties:
        artifact_id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        project_id: { type: string }
        task_id: { type: string }
        artifact_type: { type: string, enum: [CODE, DOC, DESIGN, CONFIG, REPORT, DATASET] }
        storage_uri: { type: string }
        content_hash: { type: string }
        version: { type: string }
        status: { type: string, enum: [DRAFT, IN_REVIEW, APPROVED, RELEASED, DEPRECATED] }
        parent_artifact_id: { type: string, format: uuid, nullable: true }
        agent_run_id: { type: string, nullable: true, description: "v1.4 — resolvable FK into agent_runs (clause 61)." }
        model_run_id: { type: string, nullable: true, description: "v1.4 — resolvable FK into model_runs (clause 61)." }
        reviewed_by: { type: string }
        approved_by: { type: string }
        git_commit_ref: { type: string }
        deployment_ref: { type: string, deprecated: true, description: "Superseded by /deployments and deployment_registry (clause 66); retained read-only for backward compatibility." }
        created_at: { type: string, format: date-time }
...
    ArtifactCreateRequest:
      type: object
      required: [project_id, artifact_type, storage_uri, content_hash]
      properties:
        project_id: { type: string }
        task_id: { type: string }
        agent_run_id: { type: string }
        model_run_id: { type: string }
        artifact_type: { type: string, enum: [CODE, DOC, DESIGN, CONFIG, REPORT, DATASET] }
        storage_uri: { type: string }
        content_hash: { type: string }
        version: { type: string, default: "1.0.0" }
    ArtifactLineage:
      type: object
      properties:
        artifact: { $ref: '#/components/schemas/Artifact' }
        agent_run: { type: object }
        model_run: { type: object }
        task: { $ref: '#/components/schemas/Task' }
        workflow: { $ref: '#/components/schemas/Workflow' }
        parent_chain: { type: array, items: { $ref: '#/components/schemas/Artifact' } }
...
    DeploymentCreateRequest:
      type: object
      required: [project_id, release_id, environment, strategy]
      properties:
        project_id: { type: string }
        release_id: { type: string }
        environment: { type: string }
        strategy: { type: string, enum: [standard, rolling, blue_green, canary] }
        approval_request_id: { type: string }
    Deployment:
      allOf:
        - $ref: '#/components/schemas/DeploymentCreateRequest'
        - type: object
          properties:
            deployment_id: { type: string }
            tenant_id: { type: string, format: uuid }
            status: { type: string }
            rollback_target: { type: string }
            started_at: { type: string, format: date-time }
            completed_at: { type: string, format: date-time }
    PolicyDecisionRecord:
      type: object
      properties:
        policy_decision_id: { type: string, format: uuid }
        task_id: { type: string }
        agent_id: { type: string }
        decision: { type: string, enum: [ALLOW, DENY, REQUIRE_APPROVAL, REQUIRE_ESCALATION] }
        policy_version: { type: string }
        alternatives: { type: object }
        rejection_reasons: { type: object }
        created_at: { type: string, format: date-time }
    RoutingDecisionRecord:
      type: object
      properties:
        routing_decision_id: { type: string, format: uuid }
        task_id: { type: string }
        selected_provider: { type: string }
        selected_model: { type: string }
        candidate_models: { type: object }
        reason: { type: string }
        created_at: { type: string, format: date-time }
  responses:
...
    NotFound:
      description: Resource not found (or not visible to this tenant)
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
paths:
...
  /artifacts:
    # NOTE (v1.4 integrity fix): same duplicate-key issue as
    # /agents/{agent_id}/messages above — GET and POST were declared as two
    # separate `/artifacts:` mapping keys in earlier drafts. Merged here.
    get:
      summary: List artifacts for a project
      parameters:
        - { name: project_id, in: query, required: true, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        '200':
          description: Artifact list
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Artifact' } }
    post:
      summary: Register a new artifact (immutable once status=APPROVED — clause 48)
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ArtifactCreateRequest' } } }
      responses:
        '201': { description: Artifact created, content: { application/json: { schema: { $ref: '#/components/schemas/Artifact' } } } }
  /artifacts/{artifact_id}:
    get:
      summary: Get artifact record and lineage
      parameters:
        - { name: artifact_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Artifact record
          content: { application/json: { schema: { $ref: '#/components/schemas/Artifact' } } }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /artifacts/{artifact_id}/lineage:
    get:
      summary: Resolve full lineage for an artifact (Owner Command → ... → Release — clause 23/61)
      parameters: [ { name: artifact_id, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses:
        '200': { description: Full lineage, content: { application/json: { schema: { $ref: '#/components/schemas/ArtifactLineage' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /deployments:
    post:
      summary: Create a deployment for a release (clause 66)
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/DeploymentCreateRequest' } } }
      responses:
        '201': { description: Deployment created, content: { application/json: { schema: { $ref: '#/components/schemas/Deployment' } } } }
  /deployments/{deployment_id}:
    get:
      summary: Get deployment status
      parameters: [ { name: deployment_id, in: path, required: true, schema: { type: string } } ]
      responses:
        '200': { description: Deployment record, content: { application/json: { schema: { $ref: '#/components/schemas/Deployment' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /deployments/{deployment_id}/rollback:
    post:
      summary: Roll back a deployment to its recorded rollback_target
      parameters:
        - { name: deployment_id, in: path, required: true, schema: { type: string } }
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202': { description: Rollback triggered }
  /policy-decisions:
    get:
      summary: Query structured policy decision records (clause 72)
      parameters:
        - { name: task_id, in: query, schema: { type: string } }
        - { name: agent_id, in: query, schema: { type: string } }
      responses:
        '200': { description: Policy decisions, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/PolicyDecisionRecord' } } } } }
  /routing-decisions:
    get:
      summary: Query structured routing decision records (clause 72)
      parameters:
        - { name: task_id, in: query, schema: { type: string } }
      responses:
        '200': { description: Routing decisions, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/RoutingDecisionRecord' } } } } }